FOR REAL THIS TIME

+++ Vika7/p1/measure_tlb.c
@@ -35,7 +35,7 @@ int accessMemory(uint64_t memsize, uint64_t count, uint64_t step) {
 	}
 
 
-	int current_step = 0;
+	uint64_t current_step = 0;
 	for(uint64_t i = 0; i < count; i++)
 	{
 		current_step += step;
